backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
whitenoise>=6.4.0
Pillow>=9.5.0
celery>=5.2.0
redis>=4.5.0